yoke::erased

Trait ErasedDestructor

pub trait ErasedDestructor: 'static { }

Dummy trait that lets us dyn Drop

dyn Drop isn’t legal (and doesn’t make sense since Drop is not implement on all destructible types). However, all trait objects come with a destructor, so we can just use an empty trait to get a destructor object.