Crate yoke

This crate provides Yoke, a data structure that allows one to "yoke" Cow-like borrowed data types to their backing storage, enabling one to use Cow (etc) in zero-copy deserialization with dynamic lifetimes for the borrowed data, for example caching it.

See the documentation of Yoke for more details.

Modules

trait_hack
Workarounds for adding trait bounds to yoke objects.

Structs

Yoke
A Cow-like borrowed object "yoked" to its backing data.

Traits

CloneableCart
This trait marks cart types that do not change source on cloning

Yokeable
A Yokeable type is essentially one with a covariant lifetime parameter, matched to the parameter in the trait definition. The trait allows one to cast the covariant lifetime to and from 'static.

ZeroCopyFrom
Trait for types that can be crated from a reference to a cart type C with no allocations.

Derive Macros

Yokeable
Custom derive for yoke::Yokeable,

ZeroCopyFrom
Custom derive for yoke::ZeroCopyFrom,