Trait yoke::cartable_ptr::CloneableCartablePointerLike

pub unsafe trait CloneableCartablePointerLike: CartablePointerLike { }

An object that implements CartablePointerLike that also supports cloning without changing the address of referenced data.

Safety

Implementer safety:

addref_raw must create a new owner such that an additional call to drop_raw does not create a dangling pointer

addref_raw must not change the address of any referenced data.

Object Safety

This trait is not object safe.