Trait yoke::CloneableCart

pub unsafe trait CloneableCart: Clone { }

This trait marks cart types that do not change source on cloning

This is conceptually similar to [stable_deref_trait::CloneStableDeref], however [stable_deref_trait::CloneStableDeref] is not (and should not) be implemented on Option (since it's not Deref). CloneableCart essentially is "if there is data to borrow from here, cloning the cart gives you an additional handle to the same data".