Trait yoke::ZeroCopyFrom

pub trait ZeroCopyFrom<'zcf, C: ?Sized>: 'zcf {
    fn zero_copy_from(cart: &'zcf C) -> Self;
}

Trait for types that can be created from a reference to a cart type C with no allocations.

A type can be the ZeroCopyFrom target of multiple cart types.

The intention is for ZeroCopyFrom to produce a struct from a cart with as little work as possible. Although it is technically possible to implement ZeroCopyFrom without being zero-copy (using heap allocations), doing so defeats the purpose of ZeroCopyFrom.

For example, impl ZeroCopyFrom<C> for Cow<str> should return a Cow::Borrowed pointing at data in the cart C, even if the cart is itself fully owned.

One can use the #[derive(ZeroCopyFrom)] custom derive to automatically implement this trait.

Examples

Implementing ZeroCopyFrom on a custom data struct:

use yoke::Yokeable;
use yoke::ZeroCopyFrom;
use std::borrow::Cow;

struct MyStruct<'data> {
    message: Cow<'data, str>,
}

// Reference from a borrowed version of self
impl<'zcf> ZeroCopyFrom<'zcf, MyStruct<'_>> for MyStruct<'zcf> {
    fn zero_copy_from(cart: &'zcf MyStruct<'_>) -> Self {
        MyStruct {
            message: Cow::Borrowed(&cart.message)
        }
    }
}

// Reference from a string slice directly
impl<'zcf> ZeroCopyFrom<'zcf, str> for MyStruct<'zcf> {
    fn zero_copy_from(cart: &'zcf str) -> Self {
        MyStruct {
            message: Cow::Borrowed(cart)
        }
    }
}

Required methods

fn zero_copy_from(cart: &'zcf C) -> Self

Clone the cart C into a struct that may retain references into C.